Call the correct base function in the presence of multiple virtual overloads
Now instead of passing in CxxBase as a template parameter, a configuration
struct is passed in. This will hopefully make it easier to pass in config
things in the future.

To detect the multiple virtual overload, the config struct inherits
from the config structs of the base trampolines. Each overloaded function
has the type to call in the configuration -- so if a child doesn't have
an implementation of a base overload, then the base is called instead.
This mirrors how C++ overload resolution works.